Move close behavior to view
