+++ js/PageNavView.js
@@ -2,6 +2,7 @@ import Adapt from 'core/js/adapt';
 import router from 'core/js/router';
 import ComponentView from 'core/js/views/componentView';
 import tooltips from 'core/js/tooltips';
+import logging from 'core/js/logging';
 
 class PageNavView extends ComponentView {
   initialize() {
@@ -31,12 +32,36 @@ class PageNavView extends ComponentView {
     const $target = $(event.currentTarget);
     const index = $target.data('item-index');
     const item = this.model.get('_items')[index];
+    const type = item.type;
 
+    // Close button
+    if (type === '_close') {
+      this.closeWindow();
+      return;
+    }
+
+    // Check locked or selected
     const isLocked = item._isHidden || item._isLocked;
     const isSelected = item._isCurrent;
     if (isLocked || isSelected) return;
 
+    // Navigate to the location
     this.navigateTo(item._id);
+  };
+
+  closeWindow() {
+    try {
+      const scormWrapper = require('extensions/adapt-contrib-spoor/js/scorm/wrapper');
+      if (scormWrapper) {
+        const scormWrapperInstance = scormWrapper.getInstance();
+        if (scormWrapperInstance.lmsConnected && !scormWrapperInstance.finishCalled) {
+          scormWrapperInstance.finish();
+        }
+      }
+    } catch (err) {
+      logging.warn(`Could not close window. Error: ${err}`);
+    }
+    top.window.close();
   };
 
   setupTooltips() {
